requirements.txt
# Flask y extensiones
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0

# Base de datos
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Seguridad
bcrypt==4.1.2
python-dotenv==1.0.0

# Validación
email-validator==2.1.0

# Utilidades
python-dateutil==2.8.2
Pillow==10.1.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0

# Desarrollo
black==23.12.1
flake8==7.0.0
